fastapi-back-end-test/requirements.txt
fastapi[all]
databases
alembic
pymysql
verify-email
python-jose[cryptography]
passlib[bcrypt]
redis
httpx
pytest
loguru
sqlalchemy-serializer
sqlalchemy-utils